recipes/MOABB/extra_requirements.txt
moabb
mne
orion[profet]
sklearn
torchinfo
